toolkit/components/osfile/tests/mochi/test_osfile_back.xul

<?xml version="1.0"?>
<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<window title="Testing OS.File on a chrome worker thread"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="test();">

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"/>
  <script type="application/javascript">
  <![CDATA[

let worker;

function test() {
  ok(true, "test_osfile.xul: Starting test");
  if (navigator.platform.indexOf("Win") != -1) {
    ok(true, "test_osfile.xul: Using Windows test suite");
    worker = new ChromeWorker("worker_test_osfile_win.js");
  } else {
    ok(true, "test_osfile.xul: Using Unix test suite");
    worker = new ChromeWorker("worker_test_osfile_unix.js");
  }
  SimpleTest.waitForExplicitFinish();
  ok(true, "test_osfile.xul: Chrome worker created");
  dump("MAIN: go\n");
  worker.onmessage = function(msg) {
    ok(true, "MAIN: onmessage "+JSON.stringify(msg.data));
    switch (msg.data.kind) {
    case "is":
      SimpleTest.ok(msg.data.outcome, msg.data.description +
         "( "+ msg.data.a + " ==? " + msg.data.b + ")" );
      return;
    case "isnot":
      SimpleTest.ok(msg.data.outcome, msg.data.description +
      "( "+ msg.data.a + " !=? " + msg.data.b + ")" );
         return;
    case "ok":
      SimpleTest.ok(msg.data.condition, msg.data.description);
      return;
    case "finish":
      SimpleTest.finish();
      return;
    default:
      SimpleTest.ok(false, "test_osfile.xul: wrong message "+JSON.stringify(msg.data));
      return;
    }
  };
  worker.postMessage(0);
  ok(true, "test_osfile.xul: Test in progress");
};
]]>
  </script>

  <body xmlns="http://www.w3.org/1999/xhtml">
    <p id="display"></p>
    <div id="content" style="display:none;"></div>
    <pre id="test"></pre>
  </body>
  <label id="test-result"/>
</window>
